feat: add scheduled connections to db (#31)
* add bucket methods

* Add methods to postgres

* add bucket to scheduler

* fix bucket limits

* create table

* add insert to postgres

* store schedule conenction

* remove api db fns

* fix lint

* add job id

* fix db methods

+++ src/common/device.rs
@@ -8,16 +8,16 @@ pub struct Device {
     pub ipv4: Option<String>,
     pub ipv6: Option<String>,
     pub mac: Option<String>,
-    pub factory_id: Option<i32>,
-    pub batch_id: Option<i32>,
+    pub factory_id: Option<i64>,
+    pub batch_id: Option<i64>,
 }
 
 impl Device {
     pub fn new(
         socket: SocketAddr,
         mac: Option<String>,
-        factory_id: Option<i32>,
-        batch_id: Option<i32>,
+        factory_id: Option<i64>,
+        batch_id: Option<i64>,
     ) -> Self {
         let (ipv4, ipv6) = match socket {
             SocketAddr::V4(_) => (Some(socket.ip().to_string()), None),
